requirements.txt
openai==0.27.2
streamlit==1.24.0
python-pptx==0.6.21